+++ lib/shared_map.dart
@@ -3,3 +3,4 @@ library shared_map;
 
 export 'src/shared_map_base.dart';
 export 'src/shared_map_cached.dart';
+export 'src/shared_map_extension.dart';