+++ lib/src/shared_map_generic.dart
@@ -75,6 +75,13 @@ class SharedMapGeneric<K, V> implements SharedMapSync<K, V> {
   int length() => _entries.length;
 
   @override
+  int clear() {
+    var lng = _entries.length;
+    _entries.clear();
+    return lng;
+  }
+
+  @override
   SharedMapReference sharedReference() =>
       SharedMapReferenceGeneric(id, sharedStore.sharedReference());
 
@@ -129,6 +136,9 @@ class SharedMapCacheGeneric<K, V> implements SharedMapCached<K, V> {
 
   @override
   int length({Duration? timeout, bool refresh = false}) => _sharedMap.length();
+
+  @override
+  int clear() => _sharedMap.clear();
 
   @override
   SharedMapCached<K, V> cached({Duration? timeout}) => this;